Pi/requirements.txt
# XIOT Pi Requirements
smbus2>=0.4.0
paho-mqtt>=2.0.0

